refactor(optimize): streamline optimization algorithms
Improve code quality across constrained, global, and QP optimization:

- Add type alias ConstraintFn for cleaner function signatures
- Derive Default for QpMethod instead of manual implementation
- Simplify control flow in COBYLA acceptance criteria
- Add clippy::too_many_arguments and clippy::type_complexity allows for
  complex optimization algorithms where parameter count is inherent
- Remove redundant type annotations and unnecessary casts
- Replace iterator patterns for better readability in active set QP

+++ src/optimize/global/impl_generic/shgo.rs
@@ -59,7 +59,7 @@ where
     for d in 0..dim {
         let base = PRIMES.get(d).copied().unwrap_or(113);
         let x = van_der_corput(index, base);
-        values.push(x as f64);
+        values.push(x);
     }
 
     let device = client.device();
